autoposting/tg/day_in_history/requirements.txt
# =============================================================================
# requirements.txt - ЗАВИСИМОСТИ ДЛЯ БОТА "ДЕНЬ В ИСТОРИИ"
# =============================================================================

requests==2.31.0
# 🔄 HTTP-запросы к Telegram API
# Лицензия: Apache 2.0

schedule==1.2.0
# ⏰ Планировщик ежедневных задач
# Лицензия: MIT

python-dotenv==1.0.0
# 🔐 Загрузка конфигурации из .env файла
# Лицензия: BSD-3-Clause

# =============================================================================
# Совместимость: Python 3.8+
# =============================================================================